fix(verilator): use uint64_t for no-timing harness sim_time
The --no-timing C++ harness declared 'unsigned long long sim_time', which on Linux (uint64_t = unsigned long) makes tfp->dump(sim_time++) ambiguous across the VerilatedVcdC::dump() overloads — breaking every --no-timing chapter (all spivoop* / SPI OOP) once its SV compiles. uint64_t is an exact match on all platforms; verified spivoop2 compiles and prints all expected lines. Adds error.md documenting this plus the 'Unknown verilator comment' trap, linked from CLAUDE.md Step 4f.

+++ main.py
@@ -49,7 +49,9 @@ int main(int argc, char** argv) {
     top->trace(tfp, 99);
     tfp->open("wave.vcd");
 
-    unsigned long long sim_time = 0;
+    uint64_t sim_time = 0;   // uint64_t, NOT 'unsigned long long': on Linux uint64_t is
+                             // 'unsigned long', so ull makes tfp->dump() ambiguous. See error.md
+
     while (!Verilated::gotFinish() && sim_time < 1000000ULL) {
         top->eval();
         tfp->dump(sim_time++);
